internal/plot: merge list.go in plots.go
